Use Path:: methods instead of Filesystem instance
`Filesystem->isAbsolutePath` does not check if the file exists and just make quick checks on the string.

So here the Path utility methods seem a better choice as they are platform-agnostic and allow safe operations like `::join`  handle `../../` in paths and handle full path canonicalization.

And the tests can be done on any strings, independently on the Fixtures files here.

I think this method is used before calling `new File()`, so maybe the whole thing could be replaced by some factory/provider returning an Asset or something like that?